fetch all data and rerun all code
Fix history_years in shared.py.

Fix name handling trying to read file before writing it.
