+++ reports/post_race_analysis_ju_fy_2019.txt
@@ -1,54 +1,54 @@
-Yksilöennusteen keskivirhe: 15.956 minuuttia
-Yksilöennusteen mediaanivirhe: 10.34 minuuttia
-Yksilön aikaväliennuste väärin: 9.67 %
+Yksilöennusteen keskivirhe: 15.798 minuuttia
+Yksilöennusteen mediaanivirhe: 10.282 minuuttia
+Yksilön aikaväliennuste väärin: 9.370000000000001 %
          ind_interval         ind_error        ind_interval_error      
                  mean  median      mean median               mean count
 num_runs                                                               
-0              113.78  106.82     25.30  17.37               0.08  2745
-1               96.04   85.05     16.07  11.00               0.07  1688
-2              122.64  108.87     14.75   9.46               0.01  1201
-3               82.19   73.22     12.98   8.86               0.05  1116
-4               50.77   40.66     13.12   8.91               0.16  1004
-5               50.51   43.16     12.76   8.00               0.12  1005
-6               46.89   40.33     12.71   8.87               0.16  1181
-7               45.54   39.93     11.98   8.88               0.13  1373
-8               41.95   36.58     11.02   8.28               0.13  1146
-9               34.51   37.51      6.15   5.14               0.00     5
-11              35.21   35.21      6.14   6.14               0.00     2
-Viestiennusteen keskivirhe: 39.886 minuuttia
-Viestiennusteen virheen mediaani: 23.106 minuuttia
+0              119.91  110.01     24.58  17.13               0.06  2745
+1               96.14   80.76     16.07  11.00               0.07  1688
+2              122.03  108.99     14.75   9.46               0.02  1201
+3               82.42   70.85     12.98   8.86               0.05  1116
+4               50.81   40.46     13.12   8.91               0.16  1004
+5               50.54   43.14     12.76   8.00               0.12  1005
+6               46.95   40.51     12.71   8.87               0.16  1181
+7               45.52   39.78     11.98   8.88               0.13  1373
+8               41.96   36.64     11.02   8.28               0.13  1146
+9               35.03   38.70      6.15   5.14               0.00     5
+11              35.15   35.15      6.14   6.14               0.00     2
+Viestiennusteen keskivirhe: 38.801 minuuttia
+Viestiennusteen virheen mediaani: 23.083 minuuttia
     relay_error             
            mean median count
 leg                         
-1         11.87   7.36  1897
-2         21.34  13.19  1872
-3         30.54  19.21  1837
-4         41.16  27.45  1789
-5         53.01  36.66  1747
-6         61.51  45.82  1690
-7         66.37  48.99  1634
-Viestin aikaväliennuste väärin: 11.5 %
+1         11.99   7.61  1897
+2         21.00  13.47  1872
+3         30.12  19.50  1837
+4         40.18  26.94  1789
+5         51.49  36.49  1747
+6         59.05  44.91  1690
+7         64.06  48.35  1634
+Viestin aikaväliennuste väärin: 11.3 %
     real_interval real_interval_error real_interval_error_fast real_interval_error_slow
              mean                mean                     mean                     mean
 leg                                                                                    
-1          73.601               0.054                    0.011                    0.043
-2         103.635               0.080                    0.007                    0.073
-3         144.213               0.076                    0.005                    0.071
-4         160.942               0.113                    0.003                    0.110
-5         177.787               0.161                    0.003                    0.157
-6         201.192               0.176                    0.002                    0.175
-7         225.549               0.162                    0.001                    0.160
+1          72.512               0.056                    0.012                    0.044
+2         103.569               0.082                    0.007                    0.074
+3         145.425               0.072                    0.004                    0.068
+4         163.742               0.109                    0.002                    0.107
+5         182.062               0.158                    0.004                    0.154
+6         206.354               0.172                    0.003                    0.169
+7         230.911               0.159                    0.004                    0.154
               real_interval real_interval_error real_interval_error_fast real_interval_error_slow
                        mean                mean                     mean                     mean
 mean_num_runs                                                                                    
-0.0                 208.611               0.162                    0.002                    0.161
-1.0                 220.693               0.098                    0.016                    0.082
-2.0                 200.020               0.062                    0.002                    0.060
-3.0                 175.884               0.070                    0.007                    0.063
-4.0                 142.018               0.093                    0.004                    0.089
-5.0                 120.257               0.139                    0.003                    0.135
-6.0                  96.762               0.170                    0.004                    0.166
-7.0                  76.948               0.160                    0.000                    0.160
-8.0                  57.109               0.150                    0.002                    0.147
-9.0                  35.311               0.000                    0.000                    0.000
-11.0                 31.913               0.000                    0.000                    0.000
+0.0                 215.139               0.154                    0.010                    0.144
+1.0                 226.543               0.091                    0.018                    0.072
+2.0                 204.252               0.058                    0.003                    0.055
+3.0                 177.506               0.073                    0.004                    0.069
+4.0                 143.419               0.093                    0.003                    0.090
+5.0                 120.604               0.140                    0.003                    0.137
+6.0                  97.330               0.168                    0.004                    0.164
+7.0                  77.124               0.162                    0.000                    0.162
+8.0                  57.224               0.147                    0.002                    0.145
+9.0                  36.097               0.000                    0.000                    0.000
+11.0                 31.666               0.000                    0.000                    0.000
