+++ reports/preprocess_priors_grouped_ju_fy_2019.txt
@@ -2,21 +2,21 @@ x.shape: (30010, 56)
 y.shape: (30010, 1)
 None
 <class 'sklearn.linear_model._base.LinearRegression'>: {'copy_X': True, 'fit_intercept': True, 'n_jobs': None, 'normalize': False}
-Explained variance score: 0.436
+Explained variance score: 0.437
 <class 'sklearn.ensemble._gb.GradientBoostingRegressor'>: {'alpha': 0.9, 'ccp_alpha': 0.0, 'criterion': 'friedman_mse', 'init': None, 'learning_rate': 0.1, 'loss': 'ls', 'max_depth': 3, 'max_features': None, 'max_leaf_nodes': None, 'min_impurity_decrease': 0.0, 'min_impurity_split': None, 'min_samples_leaf': 1, 'min_samples_split': 2, 'min_weight_fraction_leaf': 0.0, 'n_estimators': 900, 'n_iter_no_change': None, 'presort': 'deprecated', 'random_state': 0, 'subsample': 1.0, 'tol': 0.0001, 'validation_fraction': 0.1, 'verbose': 1, 'warm_start': False}
-Explained variance score: 0.486
+Explained variance score: 0.484
 <class 'sklearn.ensemble._gb.GradientBoostingRegressor'>: {'alpha': 0.159, 'ccp_alpha': 0.0, 'criterion': 'friedman_mse', 'init': None, 'learning_rate': 0.1, 'loss': 'quantile', 'max_depth': 3, 'max_features': None, 'max_leaf_nodes': None, 'min_impurity_decrease': 0.0, 'min_impurity_split': None, 'min_samples_leaf': 1, 'min_samples_split': 2, 'min_weight_fraction_leaf': 0.0, 'n_estimators': 450, 'n_iter_no_change': None, 'presort': 'deprecated', 'random_state': 0, 'subsample': 1.0, 'tol': 0.0001, 'validation_fraction': 0.1, 'verbose': 1, 'warm_start': False}
-Explained variance score: 0.079
+Explained variance score: 0.077
 <class 'sklearn.ensemble._gb.GradientBoostingRegressor'>: {'alpha': 0.841, 'ccp_alpha': 0.0, 'criterion': 'friedman_mse', 'init': None, 'learning_rate': 0.1, 'loss': 'quantile', 'max_depth': 3, 'max_features': None, 'max_leaf_nodes': None, 'min_impurity_decrease': 0.0, 'min_impurity_split': None, 'min_samples_leaf': 1, 'min_samples_split': 2, 'min_weight_fraction_leaf': 0.0, 'n_estimators': 450, 'n_iter_no_change': None, 'presort': 'deprecated', 'random_state': 0, 'subsample': 1.0, 'tol': 0.0001, 'validation_fraction': 0.1, 'verbose': 1, 'warm_start': False}
-Explained variance score: 0.136
-q_low_error.mean 0.1652
-q_high_error.mean 0.1587
-q_error.mean 0.3239
-q_interval.mean 3.3612
-q_interval.median 3.1719
-std.mean 1.6806
-std_correct.mean 1.1656
-abs_error_in_stds.mean 0.4332
-abs_error.mean 1.4308
-abs_error.median 0.9491
-runtime 0.9 mins
+Explained variance score: 0.115
+q_low_error.mean 0.1649
+q_high_error.mean 0.1601
+q_error.mean 0.3249
+q_interval.mean 3.3657
+q_interval.median 3.1608
+std.mean 1.6828
+std_correct.mean 1.1659
+abs_error_in_stds.mean 0.4323
+abs_error.mean 1.4283
+abs_error.median 0.942
+runtime 0.85 mins
